intake seeds
